+++ CHANGELOG.md
@@ -5,6 +5,11 @@ All notable changes to this project will be documented here.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+### Added
+- Improved touch actions on the overview menu
+
 ## [1.3.1] - 2018-10-18
 
 ### Changed