+++ CHANGELOG.md
@@ -5,6 +5,19 @@ All notable changes to this project will be documented here.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [1.4.0] - 2018-10-23
+
+### Added
+- Improved touch actions on the overview menu.
+- Added suggestions when adding a new tag to a plant.
+- It's now possible to login from the Settings view.
+
+### Changed
+- Updated list view with more details.
+
+### Fixed
+- Hiding tags on plant view now works again.
+
 ## [1.3.1] - 2018-10-18
 
 ### Changed